Sanskrit text etxraction : First level code
This is the first level code of an extraction project that aims at extracting Sanskrit text from PDFs. 
There are many files with the name MyData[i].pdf where i refers to the numbers starting 1 till the last datafile. In order to read each datafile, I have first created a list with all the names like 
list - ["MyData1.pdf", "MyData2.pdf"...]
And so, the PDF reader shall be traversing through this list, to read. I have found this method to be quick. Instead of manualy creating a list with all the file names, this method is quick and easy.